+++ network/server/node_http/http_server_js.go
@@ -3,6 +3,9 @@
 
 package node_http
 
-func (server *HttpServer) Initialize() {
+type HttpServer struct {
+}
 
+func NewHttpServer() (*HttpServer, error) {
+	return &HttpServer{}, nil
 }
